Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodel
pydantic
python-multipart
transformers
langchain
speechrecognition
gtts
google-api-python-client
pydantic[email]
google-generativeai
PyPDF2
python-dotenv
langchain-google-genai
langchain_google_vertexai
passlib[bcrypt]
python-jose
pyjwt
pytesseract 
pillow
faiss-cpu
langchain-community
openai
pygame
langchain-openai
reportlab
pypdf